chore: consolidate WebSocket server
